pyproject.toml
[project]
name = "sql-aiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-ollama>=0.3.10",
    "langgraph>=0.6.10",
    "mcp>=1.25.0",
    "mysql-connector-python>=9.4.0",
    "numpy>=2.3.4",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "streamlit>=1.50.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langchain-openai>=1.1.7",
]
